enable all examples
